# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for Telegram WebApp (TAM).",
        "title": "TAM Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/customers/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get profile of logged-in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.CustomerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update profile details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "description": "Update data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.UpdateMeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.CustomerResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customers/me/deals": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get deals where user is advertiser or manager",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Get user deals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.GetMyDealsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_wickedv43_TAM-backend_internal_common.CustomerResponse": {
            "type": "object",
            "properties": {
                "address_bounceable": {
                    "type": "string"
                },
                "address_nonbounceable": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common_constants_customer_status.CustomerStatus"
                },
                "tg_firstname": {
                    "type": "string"
                },
                "tg_id": {
                    "type": "integer"
                },
                "tg_is_premium": {
                    "type": "boolean"
                },
                "tg_language": {
                    "type": "string"
                },
                "tg_lastname": {
                    "type": "string"
                },
                "tg_picture": {
                    "type": "string"
                },
                "tg_username": {
                    "type": "string"
                },
                "ton_balance": {
                    "type": "number"
                },
                "ton_balance_locked": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_common.DealResponse": {
            "type": "object",
            "properties": {
                "advertiser_customer_id": {
                    "type": "string"
                },
                "channel_id": {
                    "type": "string"
                },
                "channel_manager_id": {
                    "type": "string"
                },
                "common_deadline": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "publication_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common_constants_deal_status.DealStatus"
                },
                "status_updated_at": {
                    "type": "string"
                },
                "target": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.DealTarget"
                },
                "target_type": {
                    "type": "string"
                },
                "ton_price": {
                    "type": "number"
                },
                "top_deadline": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common_constants_deal_type.DealType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_wickedv43_TAM-backend_internal_common.DealTarget": {
            "type": "object",
            "properties": {
                "brief_data": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.TgMsgData"
                },
                "post_data": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.TgMsgData"
                },
                "preferable_datetime": {
                    "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.PreferableDatetime"
                },
                "repost_data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "story_data": {
                    "description": "@TODO: Story and Repost",
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_common.GetMyDealsResponse": {
            "type": "object",
            "properties": {
                "advertiser_deals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.DealResponse"
                    }
                },
                "manager_deals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.DealResponse"
                    }
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_common.PreferableDatetime": {
            "type": "object",
            "properties": {
                "date_from": {
                    "type": "string"
                },
                "date_to": {
                    "type": "string"
                },
                "time_from": {
                    "type": "string"
                },
                "time_to": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_common.TgMsgData": {
            "type": "object",
            "properties": {
                "message_album_id": {
                    "type": "string"
                },
                "message_caption_above": {
                    "type": "boolean"
                },
                "message_deal_id": {
                    "type": "string"
                },
                "message_entities": {
                    "type": "array",
                    "items": {}
                },
                "message_images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_wickedv43_TAM-backend_internal_common.TgMsgMediaItem"
                    }
                },
                "message_preview_ids": {
                    "type": "array",
                    "items": {}
                },
                "message_text": {
                    "type": "string"
                }
            }
        },
        "github_com_wickedv43_TAM-backend_internal_common.TgMsgMediaItem": {
            "type": "object",
            "properties": {
                "caption_above": {
                    "type": "boolean"
                },
                "file_id": {
                    "type": "string"
                },
                "has_spoiler": {
                    "type": "boolean"
                },
                "msg_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_wickedv43_TAM-backend_internal_common.UpdateMeRequest": {
            "type": "object",
            "properties": {
                "address_bounceable": {
                    "type": "string"
                },
                "address_nonbounceable": {
                    "type": "string"
                },
                "tg_firstname": {
                    "type": "string"
                },
                "tg_language": {
                    "type": "string"
                },
                "tg_lastname": {
                    "type": "string"
                },
                "tg_picture": {
                    "type": "string"
                },
                "tg_username": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wickedv43_TAM-backend_internal_common_constants_customer_status.CustomerStatus": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "PRE_REGISTRATION",
                "REGISTERED"
            ]
        },
        "github_com_wickedv43_TAM-backend_internal_common_constants_deal_status.DealStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13
            ],
            "x-enum-varnames": [
                "DRAFT",
                "PENDING",
                "DISCUSSION",
                "IN_PROGRESS",
                "AWAITING_APPROVAL",
                "AWAITING_APPROVAL_TIME",
                "AWAITING_PUBLICATION",
                "PUBLISHED",
                "CANCELED",
                "EXPIRED",
                "TERMS_VIOLATED",
                "COMPLETED",
                "ERR_PUBLISH",
                "ERR_BAD_TARGET"
            ]
        },
        "github_com_wickedv43_TAM-backend_internal_common_constants_deal_type.DealType": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "ADVERTISING_OFFER",
                "CHANNEL_OFFER"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}